data-loader/pyproject.toml
[project]
name = "sloww-data-loader"
version = "0.1.0"
description = "Sloww AI ingestion worker (Postgres + Qdrant + GCS + embeddings)."
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
  "psycopg[binary,pool]>=3.2.0",
  "qdrant-client>=1.10.0",
  "google-cloud-storage>=2.18.0",
  "pypdf>=5.0.0",
  "python-docx>=1.1.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
  "pydantic-settings>=2.6.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.7.0",
  "mypy>=1.10.0",
]

[project.scripts]
sloww-data-loader = "data_loader.main:run"
sloww-data-loader-http = "data_loader.http_app:run"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["data_loader*"]


